add new code

+++ code/textattack/goal_functions/classification/classification_goal_function.py
@@ -59,7 +59,7 @@ class ClassificationGoalFunction(GoalFunction):
             # set of numbers corresponding to probabilities, which should add
             # up to 1. Since they are `torch.float` values, allow a small
             # error in the summation.
-            scores = torch.nn.functional.softmax(scores, dim=1)
+            scores = torch.nn.functional.softmax(scores.to(torch.float32), dim=1)
             if not ((scores.sum(dim=1) - 1).abs() < 1e-6).all():
                 raise ValueError("Model scores do not add up to 1.")
         return scores.cpu()
